requirements.txt
behave==1.2.5
fake-factory==0.5.3
parse==1.6.6
parse-type==0.3.4
pbr==1.8.1
selenium==2.53.5
six==1.10.0
stevedore==1.10.0
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.6.0
wheel==0.24.0
pyhamcrest
requests==2.9.1
ipdb
pyyaml
beautifulsoup4
python-dateutil